+++ pyproject.toml
@@ -4,11 +4,8 @@ version = "0.10.1-beta"
 description = "Asynchronous interface for peewee ORM powered by asyncio."
 authors = ["Alexey Kinev <rudy@05bit.com>", "Gorshkov Nikolay(contributor) <nogamemorebrain@gmail.com>"]
 readme = "README.md"
-packages = [
-    { include = "peewee_async.py" },
-    { include = "peewee_async_compat.py" },
-    { include = "peewee_asyncext.py" },
-]
+include = ["peewee_async_compat.py", "peewee_asyncext.py"]
+
 
 [tool.poetry.dependencies]
 python = "^3.8"
